Add support for encrypting the `sensu` data bag, for SSL and passwords
Passwords are stored in the `secrets` item of the data bag, with the
same structure as the sensu node attributes, ie - if the node attribute
for a desired encrypted value is `node.sensu.dashboard.password`, the
data bag location would be `sensu.secrets.dashboard.password`.

SSL certs stay at `sensu.ssl`, but should be uploaded via knife with
the `--secret-file`/`--secret` argument.