Have json_file resource create owner user and group if necessary
Currently, the sensu user and group are created by the package
installation.  The json_file resource creates files owned by that
user but there are no provisions in the Chef recipes for creating that
user.  That makes it hard to test recipes that create Sensu
configuration snippets without setting up a Sensu client.

Instead, the json_file resource should create the user and group as
necessary.