Test using ChefDK not Gemfiles
These test deps were super out of date and they've probably been updated multiple times over the history of this cookbook. Constantly working out a compatible set of deps that works on every user's ruby release is a nightmare left up to Chef employees. Use DK as the set of test deps. It works on every platform and is updated.

Signed-off-by: Tim Smith <tsmith@chef.io>